fix: Redirect log() output to stderr to avoid polluting return values
The log() function was outputting to stdout, which caused problems when
used inside functions that return values (like generate_cloud_init).
The log message was being captured as part of the return value, causing
file paths to be corrupted.

Changed log() to redirect output to stderr so it doesn't interfere with
function return values while still being visible in the console.

Signed-off-by: Anil Belur <askb23@gmail.com>